fix(mesh): scope N64 forced-blend texture state to material identity
A non-semi authored forced-blend face and a semi/rate-0 face could both resolve
to Blend at rate 0 while needing different alpha and material names, so the
cache key now carries the forced-blend class. Descriptor bit 9 is recorded as
the backface-culling disable.

Work produced in the parallel Codex session; committed here after verifying
both targets build and the suite runs (see the batch report for known failures).

+++ src/NeversoftMultitool/Core/Formats/Mesh/N64/N64LightRig.cs
@@ -95,7 +95,7 @@ public sealed record N64LightRig(Vector3 Ambient, Vector3 Colour, Vector3 Direct
 
             var lightPointer = BinaryPrimitives.ReadUInt32BigEndian(boot[(offset + 12)..]);
             var ambientPointer = BinaryPrimitives.ReadUInt32BigEndian(boot[(offset + 20)..]);
-            if (lightPointer != ambientPointer + 8)
+            if ((ulong)lightPointer != (ulong)ambientPointer + 8)
                 continue;
 
             return TryLocateBody(boot, offset);
